requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
boto3==1.25.3
botocore==1.28.3
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.16
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-log-request-id==2.0.0
django-phonenumber-field==7.0.0
django-queryset-csv==1.1.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
executing==1.1.1
fiscalyear==0.4.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
JSON-log-formatter==0.5.1
kombu==5.2.4
Markdown==3.4.1
matplotlib-inline==0.1.6
num2words==0.5.12
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pgeocode==0.3.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
PyPDF2==2.11.2
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.4
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
structlog==22.1.0
traitlets==5.4.0
typing_extensions==4.4.0
unicodecsv==0.14.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.9.0
